move NetworkConnection and NetworkSession to separate file
